requirements.txt
python-dotenv
requests
pytesseract
Pillow
selenium
